requirements.txt
psycopg2-binary==2.9.7
python-decouple==3.8